+++ packages/xandr_android/android/src/main/kotlin/de/thekorn/xandr/BannerViewFactory.kt
@@ -2,6 +2,7 @@ package de.thekorn.xandr
 
 import android.app.Activity
 import android.content.Context
+import de.thekorn.xandr.models.FlutterState
 import de.thekorn.xandr.models.toBannerAdViewOptions
 import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
