+++ packages/xandr_android/android/src/main/kotlin/de/thekorn/xandr/models/InterstitialAd.kt
@@ -1,4 +1,4 @@
-package de.thekorn.xandr
+package de.thekorn.xandr.models
 
 import com.appnexus.opensdk.InterstitialAdView
 import kotlinx.coroutines.CompletableDeferred
